+++ NEWS.md
@@ -1,5 +1,8 @@
 # cSEM 0.1.0:9000
 
+- `calculateVIFModeB()` and is now exported and supports all of cSEM's
+  `cSEMResults` classes.
+
 - `calculateRhoT()` and `calculateRhoC()` are now exported and support all of cSEM's    `cSEMResults` classes.
   
 - `calculateGoF()` is now exported and supports all of cSEM's `cSEMResults` classes.